refactor Profile Settings
